+++ mvc/src/main/java/nextstep/mvc/controller/tobe/AnnotationHandlerMapping.java
@@ -7,7 +7,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.handler.HandlerMapping;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.support.RequestMethod;
 import org.slf4j.Logger;
